Add BatchClient#range
Remove BatchReadOnlyTransaction#range
