+++ google-cloud-spanner/test/google/cloud/spanner/batch_read_only_transaction/read_test.rb
@@ -123,7 +123,7 @@ describe Google::Cloud::Spanner::BatchReadOnlyTransaction, :read, :mock_spanner 
     mock.expect :streaming_read, results_enum, [session_grpc.name, "my-table", ["id", "name", "active", "age", "score", "updated_at", "birthday", "avatar", "project_ids"], Google::Spanner::V1::KeySet.new(ranges: [Google::Cloud::Spanner::Convert.to_key_range([1,1]..[3,3])]), transaction: tx_selector, index: "MyTableCompositeKey", limit: nil, resume_token: nil, partition_token: nil, options: default_options]
     session.service.mocked_service = mock
 
-    lookup_range = batch_tx.range [1,1], [3,3]
+    lookup_range = Google::Cloud::Spanner::Range.new [1,1], [3,3]
     results = batch_tx.read "my-table", columns, keys: lookup_range, index: "MyTableCompositeKey"
 
     mock.verify
